크래프팅하면 인벤토리 바로 들어오고 아이템 재고가 없으면 빨간색 충족되면 파란색 불

+++ PixelCode/Source/PixelCode/Public/CreateItemData.h
@@ -78,17 +78,13 @@ USTRUCT(BlueprintType)
 
 	// the craftitem amount.
 	UPROPERTY(EditAnywhere, Category = "Recipe")
-	uint8 CraftedItemAmount;
+	//uint8 CraftedItemAmount;
+	int32 CraftedItemAmount;
  
 	// the Recipe needed to craft the item.
 	UPROPERTY(EditAnywhere, Category = "Recipe")
 	TArray<FRecipe> CraftRecipes;
  };
-
-
-
-
-
